[BOOKINGSG-4350][WK] move fns into helper file
